+++ src/main/java/lotto/service/generate/LottoManualGenerateService.java
@@ -1,4 +1,4 @@
-package lotto.service;
+package lotto.service.generate;
 
 import lotto.domain.Lotto;
 import lotto.exception.LottoException;
